+++ src/global.d.ts
@@ -1,4 +1,3 @@
-import { GamePF2e, DamageDamageContextFlag } from "foundry-pf2e";
 import { AssistantSocket } from "socket.ts";
 import { AssistantStorage } from "storage.ts";